Make incoming translation reliable across displays and restarts
The automatic incoming-chat path now detects macOS game windows, converts global game bounds into display-local capture regions, restores persisted overlay state, and keeps the ticker aligned through mount, move, minimise, and restore transitions. UI density and three-language copy were tightened for the fixed 1280x720 window, and CI now exercises both desktop platforms plus the proxy smoke suites before release.

Constraint: Game-window bounds are global desktop coordinates while capture surfaces are display-local and macOS window coordinates are logical points.

Rejected: Rely on one-time game-window events | the Home page and overlay can mount after the transition and miss the current state.

Confidence: high

Scope-risk: moderate

Directive: Keep GameWindow bounds and DisplayInfo origins in the same platform coordinate space; convert only at capture and physical-window boundaries.

Tested: npm build; proxy and analytics smoke tests; cargo fmt; clippy -D warnings; 66 Rust tests including Retina and negative-origin display geometry; 1280x720 Home, About, overlay, and picker screenshots.

Not-tested: Live Dota 2 or League of Legends OCR on a Windows host before hosted CI.

+++ src-tauri/src/incoming/ocr/macos.rs
@@ -28,8 +28,8 @@ use objc2::runtime::AnyObject;
 use objc2::AllocAnyThread;
 use objc2_core_foundation::CFRetained;
 use objc2_core_graphics::{
-    CGBitmapInfo, CGColorRenderingIntent, CGColorSpace, CGDataProvider, CGImage,
-    CGImageAlphaInfo, CGImageByteOrderInfo,
+    CGBitmapInfo, CGColorRenderingIntent, CGColorSpace, CGDataProvider, CGImage, CGImageAlphaInfo,
+    CGImageByteOrderInfo,
 };
 use objc2_foundation::{NSArray, NSDictionary, NSString};
 use objc2_vision::{
@@ -127,14 +127,13 @@ fn frame_to_cg_image(frame: &OcrFrame) -> Result<CFRetained<CGImage>, OcrError> 
     let size = owned.len();
     let info_ptr = Box::into_raw(owned) as *mut c_void;
 
-    let provider = unsafe {
-        CGDataProvider::with_data(info_ptr, data_ptr, size, Some(release_owned_bytes))
-    }
-    .ok_or_else(|| {
-        // Recover the Box so we don't leak when the OS rejects us.
-        unsafe { drop(Box::from_raw(info_ptr as *mut Vec<u8>)) };
-        OcrError::EngineInit("CGDataProviderCreateWithData returned null".into())
-    })?;
+    let provider =
+        unsafe { CGDataProvider::with_data(info_ptr, data_ptr, size, Some(release_owned_bytes)) }
+            .ok_or_else(|| {
+            // Recover the Box so we don't leak when the OS rejects us.
+            unsafe { drop(Box::from_raw(info_ptr as *mut Vec<u8>)) };
+            OcrError::EngineInit("CGDataProviderCreateWithData returned null".into())
+        })?;
 
     let color_space = CGColorSpace::new_device_rgb()
         .ok_or_else(|| OcrError::EngineInit("CGColorSpaceCreateDeviceRGB failed".into()))?;
@@ -190,9 +189,7 @@ fn run_vision(
     frame_w: f32,
     frame_h: f32,
 ) -> Result<Vec<TextLine>, OcrError> {
-    let request = unsafe {
-        VNRecognizeTextRequest::init(VNRecognizeTextRequest::alloc())
-    };
+    let request = unsafe { VNRecognizeTextRequest::init(VNRecognizeTextRequest::alloc()) };
 
     request.setRecognitionLevel(match opts.level {
         RecognitionLevel::Accurate => VNRequestTextRecognitionLevel::Accurate,
@@ -257,7 +254,7 @@ fn run_vision(
         );
         out.push(TextLine {
             text: s,
-            confidence: text.confidence() as f32,
+            confidence: text.confidence(),
             bbox: pixel_rect,
         });
     }
